Move online c128 MTP benchmark to registered jit
